+++ CyberEngine/src/BgfxFrame.cpp
@@ -11,7 +11,16 @@ BgfxFrame::~BgfxFrame(){
 }
 
 void BgfxFrame::drawRectangle(const cc::Vector2f& pos, const cc::Vector2f& size, const cc::Color& color){
-    bgfx::dbgTextPrintf(0, 0, 0x0f, "rect : %.2f %.2f %.2f %.2f %i %i %i %i", pos.x, pos.y, size.x, size.y, (int)color.red, (int)color.green, (int)color.blue, (int)color.alpha);
+    bgfx::dbgTextPrintf(0, 0, 0x0f,
+                        "rect : %.2f %.2f %.2f %.2f %i %i %i %i",
+                        static_cast<double>(pos.x),
+                        static_cast<double>(pos.y),
+                        static_cast<double>(size.x),
+                        static_cast<double>(size.y),
+                        static_cast<int>(color.red),
+                        static_cast<int>(color.green),
+                        static_cast<int>(color.blue),
+                        static_cast<int>(color.alpha));
 
 //        cc::Vector2f size2 = size / 2.f;
 //
